refactor: replace file system tree with constant and streamline file setup in tests
